12/6 separate logic into modules
