+++ public/index.html
@@ -17,20 +17,20 @@
     <link rel="stylesheet" href="./styles/styles.css" />
 
     <script src="../socket.io/socket.io.js" defer></script>
-    <script type="module" src="./scripts/chat.js" defer></script>
+    <script type="module" src="./scripts/main.js" defer></script>
   </head>
   <body>
     <section id="page1" class="active">
       <div>
         <form id="form">
           <label for="name">Choose Name</label><br />
-          <input type="text" id="socketName" name="name" /><br />
+          <input type="text" id="socketNameElement" name="name" /><br />
           <label for="room">Input Room ID</label><br />
-          <input type="text" id="roomId" name="room"  /><br />
+          <input type="text" id="roomIdElement" name="room"  /><br />
 
           <div>
             <button type="submit" id="createRoomElement" class="btn btn-primary">Create room</button>
-            <button type="submit" id="joinRoom" class="btn btn-primary" >
+            <button type="submit" id="joinRoomElement" class="btn btn-primary" >
               Join room
             </button>
           </div>
@@ -50,7 +50,7 @@
         </div>
     
       <div id="vote-area">
-      <form id="vote-form">
+      <form id="voteFormElement">
         <button class="btn btn-primary voteBtn" id="1">1</button>
         <button class="btn btn-primary voteBtn" id="2">2</button>
         <button class="btn btn-primary voteBtn" id="3">3</button>
@@ -66,7 +66,8 @@
       <!-- <button id="nextRoundButton"  class="btn btn-dark">Next Round</button> -->
       <button id="showVotesButton" value="showVotesButtonValue" class="btn btn-dark">Show Votes</button>
 
-      <a href="" onclick="changePage('page1')" class="btn btn-light" id="leave">Leave room</a>
+      <a href="" onclick="changePage('page1')" class="btn btn-light" id="leaveRoomButton">Leave room</a>
+      <!-- TODO: switch onclick to addEventListener -->
       <div>
         <p id="average"></p>
       </div>
