+++ public/scripts/util.js
@@ -1,0 +1,11 @@
+export function changePage(pageId) {
+  // Hide all sections
+  const sections = document.querySelectorAll("section");
+  sections.forEach((section) => {
+    section.classList.remove("active");
+  });
+
+  // Show the selected section
+  const selectedSection = document.getElementById(pageId);
+  selectedSection.classList.add("active");
+}
