+++ src/handlers/showVotes.js
@@ -1,0 +1,12 @@
+export default function showVotesHandler(io, socket, data) {
+  const { votesData } = data;
+  socket.on("showVotesButtonAction", (callback) => {
+    const currentRoomName = Array.from(socket.rooms)[1];
+    io.to(currentRoomName).emit(
+      "showVotesFromServer",
+      votesData[currentRoomName]
+    );
+
+    callback(votesData[currentRoomName]);
+  });
+}
